+++ libs/mops/src/thds/mops/k8s/jobs.py
@@ -20,8 +20,9 @@ _JOB_SOURCE = WatchingObjectSource(
     lambda _, __: client.BatchV1Api().list_namespaced_job,
     lambda job: job.metadata.name,  # type: ignore
     _get_job,
+    typename="Job",
 )
 
 
-def get_job(job_name: str) -> ty.Optional[client.models.V1Job]:
-    return _JOB_SOURCE.get(job_name)
+def get_job(job_name: str, namespace: str = "") -> ty.Optional[client.models.V1Job]:
+    return _JOB_SOURCE.get(job_name, namespace=namespace)