+++ libs/mops/src/thds/mops/remote/_once.py
@@ -6,7 +6,7 @@ FNone = ty.TypeVar("FNone", bound=ty.Callable[[], None])
 
 class Once:
     """Uses unique IDs to guarantee that an operation has only run
-    once in the lifetime of this object.
+    once in the lifetime of this object, and waits for it to be complete.
 
     Is a potential source of memory leaks, since each event will be
     stored until the entire Once object is disposed.
@@ -18,10 +18,11 @@ class Once:
 
     def run_once(self, run_id: ty.Hashable, f: FNone) -> None:
         needs_run = False
-        with self.lock:
-            if run_id not in self.events:
-                needs_run = True
-                self.events[run_id] = threading.Event()
+        if run_id not in self.events:
+            with self.lock:
+                if run_id not in self.events:
+                    needs_run = True
+                    self.events[run_id] = threading.Event()
         if needs_run:
             f()
             self.events[run_id].set()