+++ libs/mops/tests/test_mops/unit/remote/test_remote_file.py
@@ -1,13 +1,13 @@
 import tempfile
 
-from thds.mops.remote.remote_file import DestFile, trigger_dest_files_download
+from thds.mops.remote.remote_file import DestFile, trigger_dest_files_placeholder_write
 
 
-def test_trigger_dest_files_download(caplog):
+def test_trigger_dest_files_placeholder_write(caplog):
 
     with tempfile.NamedTemporaryFile() as f:
         df = DestFile(f.name, "foobar")  # type: ignore
-        trigger_dest_files_download([1, 2, tuple(), dict(a=df)])
+        trigger_dest_files_placeholder_write([1, 2, tuple(), dict(a=df)])
 
         assert "DestFile" in caplog.records[0].msg
         assert "empty serialization" in caplog.records[0].msg